First working version with Toarnbeike.SourceGeneration
