+++ Directory.Build.props
@@ -1,6 +1,5 @@
 <Project>
 	<PropertyGroup>
-		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<LangVersion>preview</LangVersion>
