+++ Directory.Packages.props
@@ -3,6 +3,7 @@
 		<ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
 	</PropertyGroup>
 	<ItemGroup Label="SourceGenerator">
+		<PackageVersion Include="Microsoft.CodeAnalysis.Analyzers" Version="5.3.0" />
 		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="5.0.0" />
 		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="5.0.0" />
 	</ItemGroup>
