+++ libs/SourceGeneration/Naming/NamingExtensions.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace Toarnbeike.SourceGeneration.Naming;
+
+/// <summary>
+/// Extensions on property symbols for naming conventions.
+/// </summary>
+public static class NamingExtensions
+{
+    extension(IPropertySymbol property)
+    {
+        /// <summary>
+        /// Convert the property name to a parameter name in camelCase.
+        /// </summary>
+        /// <returns>The name of the property in camel case: e.g. converts FullName to fullName.</returns>
+        public string ToParameterName() => property.Name.ToCamelCase();
+
+        /// <summary>
+        /// Convert the property name to a backing field name in camelCase with an underscore prefix.
+        /// </summary>
+        /// <returns>The name of the property in camel case, with an underscore prefix: e.g. converts FullName to _fullName.</returns>
+        public string ToBackingFieldName() => "_" + property.Name.ToCamelCase();
+    }
+
+    public static string ToCamelCase(this string name)
+    {
+        return string.IsNullOrEmpty(name) || char.IsLower(name[0])
+            ? name
+            : name.Length == 1
+                ? name.ToLower()
+                : char.ToLower(name[0]) + name.Substring(1);
+    }
+}
