+++ libs/SourceGeneration/NetStandardCompatibility/IsExternalInit.cs
@@ -1,0 +1,8 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+// ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices;
+
+// ReSharper disable once UnusedType.Global
+[ExcludeFromCodeCoverage]
+internal static class IsExternalInit;
