+++ src/Unions.SourceGenerator/Unions.SourceGenerator.csproj
@@ -15,4 +15,16 @@
 		<AdditionalFiles Include="Analyzers\AnalyzerReleases.Shipped.md" />
 		<AdditionalFiles Include="Analyzers\AnalyzerReleases.Unshipped.md" />
 	</ItemGroup>
+
+	<ItemGroup>
+		<!-- for compile-time references -->
+		<ProjectReference Include="..\..\libs\SourceGeneration\SourceGeneration.csproj" />
+	</ItemGroup>
+	
+	<ItemGroup>
+		<!-- For analyzer runtime -->
+		<ProjectReference Include="..\..\libs\SourceGeneration\SourceGeneration.csproj"
+		                  OutputItemType="Analyzer"
+		                  ReferenceOutputAssembly="false" />
+	</ItemGroup>
 </Project>
