+++ src/Unions/Unions.csproj
@@ -1,6 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk" >
 
 	<PropertyGroup>
+		<TargetFramework>net10.0</TargetFramework>
 		<RootNamespace>Toarnbeike.Unions.Generic</RootNamespace>
 	</PropertyGroup>
