+++ tests/Unions.Tests/Unions.Tests.csproj
@@ -2,6 +2,7 @@
 
 	<PropertyGroup>
 		<RootNamespace>Toarnbeike.Unions.Tests</RootNamespace>
+		<TargetFramework>net10.0</TargetFramework>
 		<OutputType>Exe</OutputType>
 	</PropertyGroup>
